+++ CHANGELOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add the `MouseX` and `MouseY` UGens.
 - Add the `PMOsc` UGen.
+- Add `Input::range` method to easily map the output of a UGen to a specific range of numbers.